requirements.txt
pySerial
crc>=7.1.0
bitstring
flask
flask-socketio
paho-mqtt
